main.py Cleanup
Removed unused imports
Moved sqlite3 commands into 'sqlite3_proof-of-concept.py' since the sqlite3 module won't import on the Linux server and I don't want to troubleshoot since we're not planning on going further down the sqlite path
Removed the hidden tag from the system and custrole commands since that made them unusable.  We'll rely on the DEV_SVID and SYS_PASS checks
